test(storage): catch every spelling of a device-folder address
The guard against logging a device folder's tree URI matched one Kotlin
spelling of the value it names, so the same address reached a log line
through a different local and the guard stayed green. A tree URI carries the
provider's document id and, on some providers, a path from the user's
storage; logcat is readable by anything holding READ_LOGS.

The matcher moves into a shared helper and seeds its tainted set from any
declaration typed Uri, which closes the six functions and the one field
that were spelled around it in a single edit rather than a guard each. The
finding pointed at the guard's own file; the cause is one layer up, in what
the helper considers a name worth following.

No production behaviour changes: nothing leaks today. The guard would simply
not have stopped the next one.
