+++ android/app/src/test/kotlin/com/vscodroid/ExtensionCallbackTest.kt
@@ -902,16 +902,21 @@ class WorkbenchUrlTest {
  * leak, `"at $url"` fails it, and `"at " + url` walks straight through while the
  * token goes to a release build's logcat, readable by anything holding READ_LOGS.
  *
- * So [TokenTaint] reads the file the way taint reads it. A `val`/`var`
+ * So [LogTaint] reads the file the way taint reads it. A `val`/`var`
  * declaration seeded from `workbenchUrl(...)`, `getConnectionToken()` or a
  * WebView's `url` is tainted; anything declared from a tainted name is tainted
  * too; and a `Logger` statement is an offender when a tainted name survives
- * removing every `redactToken(...)` from it. Statements are read whole, across
- * the lines a formatter wraps them onto, and string prose is dropped so the word
+ * removing everything that treats it. Statements are read whole, across the
+ * lines a formatter wraps them onto, and string prose is dropped so the word
  * "token" in a message is not mistaken for the variable. Interpolation,
  * concatenation and an intermediate local all read the same to it. The reader is
  * driven against fixed snippets in the cases below, so it is measured in both
  * directions rather than trusted in either.
+ *
+ * That reader is shared with `SafFolderLogCallSiteTest`, which seeds it with
+ * `Uri`-typed declarations for the other value in this file worth keeping out of
+ * logcat. A device folder printed here therefore fails the case below as well,
+ * and that file is where the remedy for one of those is written down.
  *
  * What still passes. The first of these is pinned as a case below, so the claim
  * is measured rather than promised, and the rest are the same shape as it:
@@ -921,15 +926,17 @@ class WorkbenchUrlTest {
  *    after it was declared;
  *  - a declaration split across lines, with `val x =` on one and the value on the
  *    next;
- *  - a token arriving by a route none of the three seeds name;
+ *  - a token arriving by a route none of the seeds name;
  *  - a value handed to a helper that logs it somewhere else. The reverse of that
- *    one is a false accusation rather than a miss: only `redactToken(...)` counts
- *    as sanitising, so any other wrapper is reported and has to be argued with;
+ *    one is a false accusation rather than a miss: `redactToken(...)` and the
+ *    two reductions are the whole list of what counts as treating a value, so
+ *    any other wrapper is reported and has to be argued with;
  *  - names are tracked across the whole file rather than per scope, so an
- *    unrelated local called `url` or `token` is held to the same rule. Every
- *    binding of either name in this file today holds the same tokened URL or the
- *    bridge's session token, so here that is the answer wanted; elsewhere it
- *    would be a nuisance.
+ *    unrelated local called `url`, `token` or `uri` is held to the same rule.
+ *    Every binding of those names in this file today holds the same tokened URL,
+ *    the bridge's session token or a device folder, so here that is the answer
+ *    wanted; elsewhere it would be a nuisance. `isWorkbenchUrl` names its locals
+ *    `parsed` and `hostName` for exactly this reason, and says so.
  *
  * Source reading, and the weaker layer for the usual reason: the statement is
  * inside an Activity method, and a plain JVM test can build no Activity. The
@@ -951,13 +958,15 @@ class NavigationTokenLoggingTest {
 
     @Test
     fun `no log statement prints a token-bearing value unredacted`() {
-        val offenders = TokenTaint.leaks(source()).map { "MainActivity.kt:$it" }
+        val offenders = LogTaint.leaks(source()).map { "MainActivity.kt:$it" }
 
         assertEquals(
             emptyList<String>(), offenders,
-            "the URL the WebView loads carries the connection token, and only " +
-                "Logger.d is gated on a debuggable build, so this reaches a release " +
-                "build's logcat. Print it through redactToken().",
+            "the URL the WebView loads carries the connection token, only Logger.d is " +
+                "gated on a debuggable build, so this reaches a release build's " +
+                "logcat. Print it through redactToken(). LogTaint follows a device " +
+                "folder's tree URI as well, and `SafFolderLogCallSiteTest` says what " +
+                "to do about one of those: name it by its mirror.",
         )
     }
 
@@ -968,13 +977,13 @@ class NavigationTokenLoggingTest {
         // `val safe = redactToken(url)` that nothing logs, which is exactly the
         // company a raw log statement keeps. Deleting the log statement, renaming
         // the local, or logging the URL by some other route all leave this empty.
-        val redacted = TokenTaint.redactedLogs(source())
+        val redacted = LogTaint.redactedLogs(source())
 
         assertTrue(
             redacted.isNotEmpty(),
             "no log statement in MainActivity prints a token-bearing value through " +
                 "redactToken. Either the navigation log went, or it stopped going " +
-                "through the redactor, or the seeds in TokenTaint no longer recognise " +
+                "through the redactor, or the seeds in LogTaint no longer recognise " +
                 "where the token enters the file, and in every one of those cases the " +
                 "test above is passing by looking at nothing",
         )
@@ -985,7 +994,7 @@ class NavigationTokenLoggingTest {
         // The affordance, not the symptom. Two expressions for the same URL is
         // what made the redaction a matter of discipline; one cannot drift from
         // itself. Read off the raw lines: the literal host is string prose, which
-        // is the one thing TokenTaint's reader throws away.
+        // is the one thing LogTaint's reader throws away.
         val builders = source().withIndex()
             .filterNot { (_, l) ->
                 val t = l.trimStart()
@@ -1003,7 +1012,7 @@ class NavigationTokenLoggingTest {
 
     // --- The reader itself, driven against fixed snippets. ------------------
     //
-    // Everything above points TokenTaint at one file, where the only measurable
+    // Everything above points LogTaint at one file, where the only measurable
     // outcome is "found nothing". A reader that always finds nothing passes all
     // of it. These give it sources whose answer is known.
 
@@ -1025,9 +1034,9 @@ class NavigationTokenLoggingTest {
     fun `the code as it stands reads clean`() {
         val clean = navigateSource(redactedLog)
 
-        assertEquals(emptyList<String>(), TokenTaint.leaks(clean))
+        assertEquals(emptyList<String>(), LogTaint.leaks(clean))
         assertTrue(
-            TokenTaint.redactedLogs(clean).isNotEmpty(),
+            LogTaint.redactedLogs(clean).isNotEmpty(),
             "the reader did not recognise the redacted log statement it is built around",
         )
     }
@@ -1052,7 +1061,7 @@ class NavigationTokenLoggingTest {
 
         assertLeaks(defeat)
         assertEquals(
-            emptyList<String>(), TokenTaint.redactedLogs(defeat),
+            emptyList<String>(), LogTaint.redactedLogs(defeat),
             "a `val safeUrl = redactToken(url)` that nothing logs must not answer for " +
                 "the log statement; that is how a control gets satisfied by a line it " +
                 "has nothing to do with",
@@ -1090,9 +1099,9 @@ class NavigationTokenLoggingTest {
         )
 
         assertLeaks(wrappedRaw)
-        assertEquals(emptyList<String>(), TokenTaint.leaks(wrappedRedacted))
+        assertEquals(emptyList<String>(), LogTaint.leaks(wrappedRedacted))
         assertTrue(
-            TokenTaint.redactedLogs(wrappedRedacted).isNotEmpty(),
+            LogTaint.redactedLogs(wrappedRedacted).isNotEmpty(),
             "a redacted log statement stopped counting as one once it was wrapped",
         )
     }
@@ -1109,7 +1118,7 @@ class NavigationTokenLoggingTest {
             redactedLog,
         )
 
-        assertEquals(emptyList<String>(), TokenTaint.leaks(prose))
+        assertEquals(emptyList<String>(), LogTaint.leaks(prose))
     }
 
     @Test
@@ -1130,207 +1139,20 @@ class NavigationTokenLoggingTest {
             """        Logger.w(tag, message.toString())""",
         )
 
-        assertEquals(emptyList<String>(), TokenTaint.leaks(laundered))
+        assertEquals(emptyList<String>(), LogTaint.leaks(laundered))
         assertTrue(
-            TokenTaint.redactedLogs(laundered).isNotEmpty(),
+            LogTaint.redactedLogs(laundered).isNotEmpty(),
             "the control has to be satisfied here, or this case is not the gap it claims",
         )
     }
 
     private fun assertLeaks(source: List<String>) {
         assertTrue(
-            TokenTaint.leaks(source).isNotEmpty(),
+            LogTaint.leaks(source).isNotEmpty(),
             "a token-bearing value reaches Logger here and the reader did not see it:\n" +
                 source.joinToString("\n"),
         )
     }
-}
-
-/**
- * Which `Logger` statements in a Kotlin source hand on a value carrying the
- * connection token.
- *
- * Kept apart from the test that uses it because it is the part with behaviour of
- * its own: the cases in [NavigationTokenLoggingTest] drive it against sources
- * whose answer is known, which is only possible while it takes lines rather than
- * a filename.
- *
- * Three passes over the text, none of them a Kotlin parser and none pretending to
- * be. [codeView] drops string prose while keeping what a `$` interpolation names,
- * so a message can talk about a token without being one. [statements] gathers a
- * `Logger` call across however many lines it was wrapped onto, by counting
- * parentheses on that same prose-free view, which is what keeps a `" ("` in a
- * message from unbalancing the count. [taintedNames] walks declarations to a
- * fixpoint from the three places the token enters.
- *
- * Its blind spots are the docstring on [NavigationTokenLoggingTest], and one more
- * that belongs to the text scanning rather than to the design: a double quote
- * written inside an interpolation inside a string, `trim('"')` is one, and
- * MainActivity has it, ends the literal early, because handling it properly
- * means recursing into interpolations. The damage is bounded to that one
- * statement, and it cannot hide an interpolated name: [leaks] asks twice, once of
- * the prose-free view and once of the raw text, and only the first is affected.
- */
-private object TokenTaint {
-
-    /** Where a value carrying the connection token enters a file. */
-    private val SOURCES = listOf(
-        Regex("""\bworkbenchUrl\("""),
-        Regex("""\bgetConnectionToken\("""),
-        Regex("""\b(?:wv|webView)\??\.url\b"""),
-    )
-
-    private val DECLARATION = Regex("""\b(?:val|var)\s+([A-Za-z_]\w*)\s*(?::[^=]*?)?=(.*)""")
-    private val INTERPOLATION = Regex("""\$\{[^}]*}|\$[A-Za-z_]\w*""")
-    private val IDENTIFIER = Regex("""[A-Za-z_]\w*""")
-
-    /** Statements printing a tainted value with nothing hiding it. */
-    fun leaks(source: List<String>): List<String> {
-        val tainted = taintedNames(source)
-        return statements(source).filter { st ->
-            val code = stripRedacted(st.code)
-            val raw = stripRedacted(st.raw)
-            tainted.any { mentions(code, it) } ||
-                INTERPOLATION.findAll(raw).any { hole ->
-                    IDENTIFIER.findAll(hole.value).any { it.value in tainted }
-                }
-        }.map { "${it.line}: ${it.raw}" }
-    }
-
-    /** Statements printing a tainted value through the redactor. */
-    fun redactedLogs(source: List<String>): List<String> {
-        val tainted = taintedNames(source)
-        return statements(source).filter { st ->
-            redactedArguments(st.raw).any { arg -> tainted.any { mentions(arg, it) } }
-        }.map { "${it.line}: ${it.raw}" }
-    }
-
-    private class Statement(val line: Int, val raw: String, val code: String)
-
-    private fun isComment(line: String): Boolean {
-        val t = line.trimStart()
-        return t.startsWith("//") || t.startsWith("*") || t.startsWith("/*")
-    }
-
-    private fun mentions(text: String, name: String): Boolean =
-        Regex("""\b${Regex.escape(name)}\b""").containsMatchIn(text)
-
-    /**
-     * The line with its string prose removed and its interpolations kept.
-     *
-     * An identifier inside a literal can only be referred to through `$`, so what
-     * is left after this is every position where a name means a value.
-     */
-    private fun codeView(line: String): String {
-        val out = StringBuilder()
-        var i = 0
-        while (i < line.length) {
-            when {
-                line.startsWith("\"\"\"", i) -> {
-                    val end = line.indexOf("\"\"\"", i + 3)
-                    val inner = if (end < 0) line.substring(i + 3) else line.substring(i + 3, end)
-                    out.append(' ').append(interpolations(inner)).append(' ')
-                    i = if (end < 0) line.length else end + 3
-                }
-                line[i] == '"' -> {
-                    var j = i + 1
-                    while (j < line.length && !(line[j] == '"' && line[j - 1] != '\\')) j++
-                    val inner = line.substring(i + 1, j.coerceAtMost(line.length))
-                    out.append(' ').append(interpolations(inner)).append(' ')
-                    i = j + 1
-                }
-                else -> out.append(line[i++])
-            }
-        }
-        return out.toString()
-    }
-
-    private fun interpolations(inner: String): String =
-        INTERPOLATION.findAll(inner).joinToString(" ") { it.value }
-
-    /** Every `Logger` call, gathered across the lines it was wrapped onto. */
-    private fun statements(source: List<String>): List<Statement> {
-        val code = source.map { if (isComment(it)) "" else codeView(it) }
-        val out = mutableListOf<Statement>()
-        var i = 0
-        while (i < source.size) {
-            if (!code[i].contains("Logger.")) {
-                i++
-                continue
-            }
-            val raw = StringBuilder()
-            val whole = StringBuilder()
-            var depth = 0
-            var j = i
-            while (j < source.size && j - i < 12) {
-                raw.append(if (isComment(source[j])) "" else source[j].trim()).append(' ')
-                whole.append(code[j]).append(' ')
-                depth += code[j].count { it == '(' } - code[j].count { it == ')' }
-                if (depth <= 0) break
-                j++
-            }
-            out += Statement(i + 1, raw.toString().trim(), whole.toString())
-            i = j + 1
-        }
-        return out
-    }
-
-    /** Names holding a value that came, however indirectly, from a [SOURCES] hit. */
-    private fun taintedNames(source: List<String>): Set<String> {
-        val code = source.filterNot(::isComment).map(::codeView)
-        val names = linkedSetOf<String>()
-        // Declarations are walked to a fixpoint rather than once in file order, so
-        // that a chain assigned in the other order is still followed.
-        repeat(3) {
-            for (line in code) {
-                for (m in DECLARATION.findAll(line)) {
-                    val name = m.groupValues[1]
-                    val initialiser = m.groupValues[2]
-                    val seeded = SOURCES.any { it.containsMatchIn(initialiser) }
-                    val derived = names.toList()
-                        .any { mentions(stripRedacted(initialiser), it) }
-                    if (seeded || derived) names += name
-                }
-            }
-        }
-        return names
-    }
-
-    /** The spans of every `redactToken(...)` call, argument list included. */
-    private fun redactSpans(text: String): List<IntRange> {
-        val spans = mutableListOf<IntRange>()
-        var from = 0
-        while (true) {
-            val at = text.indexOf("redactToken(", from)
-            if (at < 0) return spans
-            var depth = 0
-            var i = at + "redactToken".length
-            var end = -1
-            while (i < text.length) {
-                if (text[i] == '(') depth++
-                else if (text[i] == ')' && --depth == 0) {
-                    end = i
-                    break
-                }
-                i++
-            }
-            if (end < 0) {
-                spans += at..text.lastIndex
-                return spans
-            }
-            spans += at..end
-            from = end + 1
-        }
-    }
-
-    private fun stripRedacted(text: String): String {
-        var out = text
-        for (span in redactSpans(text).reversed()) out = out.removeRange(span)
-        return out
-    }
-
-    private fun redactedArguments(text: String): List<String> =
-        redactSpans(text).map { text.substring(it).substringAfter('(').removeSuffix(")") }
 }
 
 /**
